Fix osutils bug where it returns wrong exit code
