+++ internal/osutils/osutils.go
@@ -10,26 +10,8 @@ import (
 
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/fileutils"
-	"github.com/ActiveState/cli/internal/logging"
 	"github.com/ActiveState/cli/internal/multilog"
-	"github.com/ActiveState/cli/internal/rollbar"
 )
-
-// CmdExitCode returns the exit code of a command in a platform agnostic way
-// taken from https://www.reddit.com/r/golang/comments/1hvvnn/any_better_way_to_do_a_crossplatform_exec_and/caytqvr/
-func CmdExitCode(cmd *exec.Cmd) (code int) {
-	defer func() {
-		if r := recover(); r != nil {
-			multilog.Log(logging.ErrorNoStacktrace, rollbar.Error)("Could not get exit code, so returning 1 instead (this is non-fatal, but should be resolved), actual error: %v", r)
-			code = 128
-		}
-	}()
-
-	type Status interface {
-		ExitStatus() int
-	}
-	return cmd.ProcessState.Sys().(Status).ExitStatus()
-}
 
 // CmdString returns a human-readable description of c.
 // This is a copy of the Go 1.13 (cmd.String) function
