+++ internal/osutils/osutils_lin_mac.go
@@ -1,0 +1,13 @@
+//go:build !windows
+// +build !windows
+
+package osutils
+
+import (
+	"os/exec"
+)
+
+// CmdExitCode returns the exit code of a command
+func CmdExitCode(cmd *exec.Cmd) (code int) {
+	return cmd.ProcessState.ExitCode()
+}
