+++ internal/osutils/osutils_windows.go
@@ -1,0 +1,17 @@
+package osutils
+
+// CmdExitCode returns the exit code of a command in a platform agnostic way
+// taken from https://www.reddit.com/r/golang/comments/1hvvnn/any_better_way_to_do_a_crossplatform_exec_and/caytqvr/
+func CmdExitCode(cmd *exec.Cmd) (code int) {
+	defer func() {
+		if r := recover(); r != nil {
+			multilog.Log(logging.ErrorNoStacktrace, rollbar.Error)("Could not get exit code, so returning 1 instead (this is non-fatal, but should be resolved), actual error: %v", r)
+			code = 128
+		}
+	}()
+
+	type Status interface {
+		ExitStatus() int
+	}
+	return cmd.ProcessState.Sys().(Status).ExitStatus()
+}
